Stage 4.8.4c: widen pointer locals + draw-helper public API
Six print/draw helpers stored frame buffer or tile addresses in
int32_t locals, function parameters, or both — silently truncating
on x86_64. Fixed in two layers:

Internal locals (no API change):
  - print4x6 / print6x8: x/y parameters were reused as framebuffer
    cursors. Introduced intptr_t row/col locals so the parameters
    keep their int32_t coordinate semantics.
  - drawline2dclip:  int32_t p → intptr_t p
  - drawspherefill:  int32_t p,sy2 — split sy2 (still int row index
    for ftol/range-check) and a new intptr_t pend (the row-end
    address it was reused as).
  - drawtile:        int32_t p,j → intptr_t p,j (j carries the
    tf-derived source-tile row); branch-A `int32_t plc` → intptr_t.
    Branch-B `plc` is a u-coordinate, stays int32_t.

Public API widening (tile / picture base pointers):
  - drawtile(intptr_t tf, ...)   was int32_t — caller (oracle.c)
                                  had been working around it with
                                  `(int32_t)(intptr_t)g_tile`,
                                  simplified to `(intptr_t)g_tile`.
  - drawpicinquad(intptr_t rpic, ..., intptr_t wpic, ...)
                                  read + write pic base pointers.
  - drawpolyquad(intptr_t rpic, ...)
                                  read pic base pointer.

Header voxlap5.h: matching signatures. Bit-identical on Windows
MSVC x86 (intptr_t == int32_t). The original Ken Silverman code in
game/ already used `long` (64-bit on Linux) for these args — the
fork narrowed them to int32_t somewhere and this restores the
intent for x64 toolchains.

Linux GCC/Clang x86_64: int-to-pointer-cast warnings 110 → 69.
Oracle still SIGSEGVs — remaining truncations live in kplib image
renderers (4.8.4d) and voxlap5.c kv6/sprite paths (4.8.4e).

+++ tests/oracle/oracle.c
@@ -367,7 +367,7 @@ int main(void) {
 		switch (poses[i].tile) {
 		case 1:
 			/* 1× zoom, alpha disabled (black==white triggers ignore-alpha). */
-			drawtile((int32_t)(intptr_t)g_tile, TILE_SIZE * 4,
+			drawtile((intptr_t)g_tile, TILE_SIZE * 4,
 			         TILE_SIZE, TILE_SIZE, 8 << 16, 8 << 16,
 			         320 << 16, 240 << 16,
 			         1 << 16, 1 << 16,
@@ -375,7 +375,7 @@ int main(void) {
 			break;
 		case 2:
 			/* 0.5× zoom (xz=yz=32768) — fast 2×2 averaging path. */
-			drawtile((int32_t)(intptr_t)g_tile, TILE_SIZE * 4,
+			drawtile((intptr_t)g_tile, TILE_SIZE * 4,
 			         TILE_SIZE, TILE_SIZE, 8 << 16, 8 << 16,
 			         320 << 16, 240 << 16,
 			         32768, 32768,
@@ -385,7 +385,7 @@ int main(void) {
 			/* 1.5× zoom + alpha modulate-and-blend path. black/white
 			 * differ in alpha (0x40 vs 0xc0), so the alpha branch fires;
 			 * the 0x40-vs-0xc0 channel spread tints the tile cyan-ward. */
-			drawtile((int32_t)(intptr_t)g_tile, TILE_SIZE * 4,
+			drawtile((intptr_t)g_tile, TILE_SIZE * 4,
 			         TILE_SIZE, TILE_SIZE, 8 << 16, 8 << 16,
 			         320 << 16, 240 << 16,
 			         (3 << 16) / 2, (3 << 16) / 2,
